requirements.txt
asgiref==3.4.1
astroid==2.8.4
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.9
djangorestframework==3.12.4
drf-yasg==1.20.0
flake8==4.0.1
idna==3.3
inflection==0.5.1
isort==5.10.0
itypes==1.2.0
Jinja2==3.0.2
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.2
platformdirs==2.4.0
psycopg2==2.9.1
pycodestyle==2.8.0
pyflakes==2.4.0
pylint==2.11.1
pyparsing==2.4.7
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
wrapt==1.13.3
